ARUHA-591: refactored subscription service and controller;
